+++ tests/Test/Proto/Parse/Option.hs
@@ -21,8 +21,8 @@ import qualified Proto3.Suite.DotProto.Parsing as Proto3
 import Proto3.Suite.DotProto.Rendering () -- orphan Pretty DotProtoIdentifier
 
 tests :: TestTree
-tests = 
-  testGroup 
+tests =
+  testGroup
     "Test.Proto.Parse.Option"
     [ testProperty "Unqualified Option Identifier" propParseName
     , testProperty "Qualified Option Identifier" propParseQName
@@ -48,28 +48,28 @@ propParseQName = property $ do
 
 --------------------------------------------------------------------------------
 
-testsOptionKw :: TestTree 
-testsOptionKw = 
+testsOptionKw :: TestTree
+testsOptionKw =
   testGroup
     "Test.Proto.Parse.Option.Keyword"
     [ testProperty "Keyword 'Option'" propParseOptionKw
     , testProperty "Keyword Malformed" propParseOptionKwMalformed
     ]
 
-propParseOptionKw :: Property 
-propParseOptionKw = property $ do 
+propParseOptionKw :: Property
+propParseOptionKw = property $ do
   str <- forAll Gen.optionKw
-  case runParseTest Proto3.pOptionKw str of 
+  case runParseTest Proto3.pOptionKw str of
     Left err -> Hedgehog.footnoteShow err >> Hedgehog.failure
     Right () -> Hedgehog.success
 
--- | Ensure the parser handling the keyword "option" must be followed by a 
+-- | Ensure the parser handling the keyword "option" must be followed by a
 -- non-alphanumeric, otherwise the parser should fail.
 propParseOptionKwMalformed :: Property
-propParseOptionKwMalformed = property $ do 
+propParseOptionKwMalformed = property $ do
   chr <- forAll Gen.ascii
   let result :: Either ParseError ()
       result = runParseTest Proto3.pOptionKw ("option" ++ [chr])
-   in if Char.isAlphaNum chr 
+   in if Char.isAlphaNum chr
         then Hedgehog.assert (isLeft result)
         else result === Right ()